Fix: move constants
