+++ packages/ui/helpers/constants/mails.js
